+++ sentry-ruby/lib/sentry/net/http.rb
@@ -50,14 +50,15 @@ module Sentry
         return if from_sentry_sdk?
 
         request_info = extract_request_info(req)
+        request_info[:body] = req.body if Sentry.configuration.send_default_pii
+
         crumb = Sentry::Breadcrumb.new(
           level: :info,
           category: OP_NAME,
           type: :info,
           data: {
-            method: request_info[:method],
-            url: request_info[:url],
-            status: res.code.to_i
+            status: res.code.to_i,
+            **request_info
           }
         )
         Sentry.add_breadcrumb(crumb)
@@ -93,8 +94,12 @@ module Sentry
 
       def extract_request_info(req)
         uri = req.uri || URI.parse("#{use_ssl? ? 'https' : 'http'}://#{address}#{req.path}")
-        url = "#{uri.scheme}://#{uri.host}#{uri.path}" rescue uri.to_s
+        url = "#{uri.scheme}://#{uri.host}#{uri.path}"
+        url = "#{url}?#{uri.query}" if uri.query && Sentry.configuration.send_default_pii
+
         { method: req.method, url: url }
+      rescue
+        { method: req.method, url: url.to_s }
       end
     end
   end